#31 Improve frontend model files structure
