--- ninja-fe/src/model/api/Message.ts
@@ -1,15 +1,0 @@
-export interface Message {
-  content?: string;
-  type: MessageType;
-  status?: MessageStatus;
-}
-
-export enum MessageStatus {
-  LOADING = "LOADING",
-  ERROR = "ERROR",
-}
-
-export enum MessageType {
-  USER = "USER",
-  ASSISTANT = "ASSISTANT",
-}
