+++ ninja-fe/src/model/api/chat/ChatRequest.ts
