+++ ninja-fe/src/model/api/chat/ChatResponse.ts
