+++ ninja-fe/src/model/api/conversation/Conversation.ts
@@ -1,0 +1,6 @@
+import { Message } from "../message/Message.ts";
+import { ConversationSummary } from "./ConversationSummary.ts";
+
+export interface Conversation extends ConversationSummary {
+  messages: Message[];
+}
