+++ ninja-fe/src/model/api/conversation/ConversationCriteria.ts
@@ -1,0 +1,12 @@
+export interface ConversationCriteria {
+  globalSearch?: string;
+  minMessageCount?: number;
+  maxMessageCount?: number;
+  minCreatedAt?: string;
+  maxCreatedAt?: string;
+  summary?: string;
+  userFullName?: string;
+  deleted?: boolean;
+
+  getUrl(): string;
+}
