+++ ninja-fe/src/model/api/conversation/ConversationSummary.ts
@@ -1,0 +1,5 @@
+export interface ConversationSummary {
+  id: string;
+  summary: string;
+  createdAt: string;
+}
