+++ ninja-fe/src/model/api/message/Message.ts
@@ -1,0 +1,8 @@
+import { MessageStatus } from "./MessageStatus.ts";
+import { MessageType } from "./MessageType.ts";
+
+export interface Message {
+  content?: string;
+  type: MessageType;
+  status?: MessageStatus;
+}
