+++ ninja-fe/src/model/api/message/MessageStatus.ts
@@ -1,0 +1,4 @@
+export enum MessageStatus {
+  LOADING = "LOADING",
+  ERROR = "ERROR",
+}
