+++ ninja-fe/src/model/api/message/MessageType.ts
@@ -1,0 +1,4 @@
+export enum MessageType {
+  USER = "USER",
+  ASSISTANT = "ASSISTANT",
+}
