+++ ninja-fe/src/model/api/page/PageRequest.ts
@@ -1,0 +1,13 @@
+export interface PageRequest {
+  page?: number;
+  size?: number;
+  sortBy?: string[];
+  sortDir?: SortDir[];
+
+  getUrl(): string;
+}
+
+export enum SortDir {
+  ASC = "asc",
+  DESC = "desc",
+}
