+++ ninja-fe/src/model/api/patch/PatchOperation.ts
@@ -1,0 +1,5 @@
+export interface PatchOperation {
+  op: string;
+  path: string;
+  value?: unknown;
+}
