+++ ninja-fe/src/model/api/user/Role.ts
@@ -1,0 +1,5 @@
+export enum Role {
+  User = "User",
+  Mod = "Mod",
+  Admin = "Admin",
+}
