+++ ninja-fe/src/model/api/user/User.ts
@@ -1,13 +1,9 @@
+import { Role } from "./Role.ts";
+
 export interface User {
   username: string;
   firstname: string;
   lastname: string;
   email: string;
   role: Role;
-}
-
-export enum Role {
-  User = "User",
-  Mod = "Mod",
-  Admin = "Admin",
 }
